Added a new command

+++ cogs/Main.py
@@ -104,7 +104,7 @@ class Main:
 		if not msg_id:
 			return await ctx.send(content = error_string + ' **Please specify a message ID to quote.**')
 
-		if ctx.guild.id in del_commands and ctx.guild.me.permissions_in(ctx.channel).manage_messages:
+		if ctx.guild and ctx.guild.id in del_commands and ctx.guild.me.permissions_in(ctx.channel).manage_messages:
 			await ctx.message.delete()
 
 		message = None
@@ -133,6 +133,24 @@ class Main:
 				await ctx.send(content = '**' + ctx.author.display_name + '\'s reply:**\n' + reply.replace('@everyone', '@еveryone').replace('@here', '@hеre'))
 		else:
 			await ctx.send(content = error_string + ' **Could not find the specified message.**')
+
+	@commands.command(aliases = ['qp'])
+	async def quotepart(self, ctx, part_msg, *, reply = None):
+		if ctx.guild and ctx.guild.id in del_commands and ctx.guild.me.permissions_in(ctx.channel).manage_messages:
+			await ctx.message.delete()
+
+		message = None
+		async with ctx.typing():
+			async for msg in ctx.channel.history(limit = 1000, before = ctx.message):
+				if part_msg.lower() in msg.content.lower():
+					message = msg
+					break
+
+			if message:
+				message.content = part_msg
+				await ctx.send(embed = quote_embed(ctx.channel, message, ctx.author))
+			else:
+				await ctx.send(content = error_string + ' **Could not find the specified message.**')
 
 	@commands.command()
 	async def prefix(self, ctx, *, prefix = None):
